Request one specific the compound information in FoodB

Request one specific the compound information in FoodB

request_foodb_compound(
  url = "https://foodb.ca/compounds",
  compound_id = "FDB000004",
  return_form = c("list", "data.frame")
)

Arguments

url

Default is "https://foodb.ca/compounds".

compound_id

compound id. For example, FDB000004.

return_form

data.frame or list.

Value

A data frame or list.

Author

Xiaotao Shen shenxt1990@outlook.com

Examples

x = request_foodb_compound( compound_id = "FDB000004", return_form = "list")
x
#> $version
#> [1] "1.0"
#> 
#> $creation_date
#> [1] "2010-04-08 22:04:24 UTC"
#> 
#> $update_date
#> [1] "2020-02-24 19:10:06 UTC"
#> 
#> $accession
#> [1] "FDB000004"
#> 
#> $name
#> [1] "Cyanidin 3-(6''-acetyl-galactoside)"
#> 
#> $description
#> [1] "Constituent of the leaves of Nymphaea alba [CCD]. Cyanidin 3-(6''-acetyl-galactoside) is found in lowbush blueberry and highbush blueberry."
#> 
#> $synonyms
#> $synonyms$synonym
#> [1] "Cyanidin 3-(6-acetylgalactoside)"
#> 
#> $synonyms$synonym
#> [1] "Cyanidin 3-O-(6''-acetyl-galactoside)"
#> 
#> 
#> $chemical_formula
#> [1] "C23H23O12"
#> 
#> $average_molecular_weight
#> [1] "491.4215"
#> 
#> $monisotopic_moleculate_weight
#> [1] "491.1189512"
#> 
#> $iupac_name
#> [1] "3-{[(2S,3R,4S,5R,6R)-6-[(acetyloxy)methyl]-3,4,5-trihydroxyoxan-2-yl]oxy}-2-(3,4-dihydroxyphenyl)-5,7-dihydroxy-1lambda4-chromen-1-ylium"
#> 
#> $traditional_iupac
#> [1] "3-{[(2S,3R,4S,5R,6R)-6-[(acetyloxy)methyl]-3,4,5-trihydroxyoxan-2-yl]oxy}-2-(3,4-dihydroxyphenyl)-5,7-dihydroxy-1lambda4-chromen-1-ylium"
#> 
#> $cas_registry_number
#> [1] "350602-26-5"
#> 
#> $smiles
#> [1] "[H][C@]1(COC(C)=O)O[C@@]([H])(OC2=CC3=C(O)C=C(O)C=C3[O+]=C2C2=CC=C(O)C(O)=C2)[C@]([H])(O)[C@@]([H])(O)[C@@]1([H])O"
#> 
#> $inchi
#> [1] "InChI=1S/C23H22O12/c1-9(24)32-8-18-19(29)20(30)21(31)23(35-18)34-17-7-12-14(27)5-11(25)6-16(12)33-22(17)10-2-3-13(26)15(28)4-10/h2-7,18-21,23,29-31H,8H2,1H3,(H3-,25,26,27,28)/p+1/t18-,19+,20+,21-,23-/m1/s1"
#> 
#> $inchikey
#> [1] "HBXXDBKJLPLXPR-DLBZZEGUSA-O"
#> 
#> $taxonomy
#> $taxonomy$description
#> [1] "belongs to the class of organic compounds known as anthocyanidin-3-o-glycosides. These are phenolic compounds containing one anthocyanidin moiety which is O-glycosidically linked to a carbohydrate moiety at the C3-position."
#> 
#> $taxonomy$direct_parent
#> [1] "Anthocyanidin-3-O-glycosides"
#> 
#> $taxonomy$kingdom
#> [1] "Organic compounds"
#> 
#> $taxonomy$super_class
#> [1] "Phenylpropanoids and polyketides"
#> 
#> $taxonomy$class
#> [1] "Flavonoids"
#> 
#> $taxonomy$sub_class
#> [1] "Flavonoid glycosides"
#> 
#> $taxonomy$molecular_framework
#> [1] "Aromatic heteropolycyclic compounds"
#> 
#> $taxonomy$alternative_parents
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "1-benzopyrans"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "1-hydroxy-2-unsubstituted benzenoids"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "1-hydroxy-4-unsubstituted benzenoids"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "3'-hydroxyflavonoids"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "4'-hydroxyflavonoids"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "5-hydroxyflavonoids"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "7-hydroxyflavonoids"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Acetals"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Anthocyanidins"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Benzene and substituted derivatives"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Carbonyl compounds"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Carboxylic acid esters"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Catechols"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Flavonoid-3-O-glycosides"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Heteroaromatic compounds"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Hydrocarbon derivatives"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Monocarboxylic acids and derivatives"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Monosaccharides"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "O-glycosyl compounds"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Organic cations"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Organic oxides"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Oxacyclic compounds"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Oxanes"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Polyols"
#> 
#> $taxonomy$alternative_parents$alternative_parent
#> [1] "Secondary alcohols"
#> 
#> 
#> $taxonomy$substituents
#> $taxonomy$substituents$substituent
#> [1] "1-benzopyran"
#> 
#> $taxonomy$substituents$substituent
#> [1] "1-hydroxy-2-unsubstituted benzenoid"
#> 
#> $taxonomy$substituents$substituent
#> [1] "1-hydroxy-4-unsubstituted benzenoid"
#> 
#> $taxonomy$substituents$substituent
#> [1] "3'-hydroxyflavonoid"
#> 
#> $taxonomy$substituents$substituent
#> [1] "4'-hydroxyflavonoid"
#> 
#> $taxonomy$substituents$substituent
#> [1] "5-hydroxyflavonoid"
#> 
#> $taxonomy$substituents$substituent
#> [1] "7-hydroxyflavonoid"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Acetal"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Alcohol"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Anthocyanidin"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Anthocyanidin-3-o-glycoside"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Aromatic heteropolycyclic compound"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Benzenoid"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Benzopyran"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Carbonyl group"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Carboxylic acid derivative"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Carboxylic acid ester"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Catechol"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Flavonoid-3-o-glycoside"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Glycosyl compound"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Heteroaromatic compound"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Hydrocarbon derivative"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Hydroxyflavonoid"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Monocarboxylic acid or derivatives"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Monocyclic benzene moiety"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Monosaccharide"
#> 
#> $taxonomy$substituents$substituent
#> [1] "O-glycosyl compound"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Organic cation"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Organic oxide"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Organic oxygen compound"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Organoheterocyclic compound"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Organooxygen compound"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Oxacycle"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Oxane"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Phenol"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Polyol"
#> 
#> $taxonomy$substituents$substituent
#> [1] "Secondary alcohol"
#> 
#> 
#> $taxonomy$external_descriptors
#> NULL
#> 
#> 
#> $state
#> NULL
#> 
#> $predicted_properties
#>        property property property      
#> kind   "logp"   "logs"   "solubility"  
#> value  "1.14"   "-3.11"  "4.12e-01 g/l"
#> source "ALOGPS" "ALOGPS" "ALOGPS"      
#> 
#> $experimental_properties
#> NULL
#> 
#> $property
#> $property$kind
#> [1] "logp"
#> 
#> $property$value
#> [1] "0.82"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "pka_strongest_acidic"
#> 
#> $property$value
#> [1] "6.39"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "pka_strongest_basic"
#> 
#> $property$value
#> [1] "-3.6"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "iupac"
#> 
#> $property$value
#> [1] "3-{[(2S,3R,4S,5R,6R)-6-[(acetyloxy)methyl]-3,4,5-trihydroxyoxan-2-yl]oxy}-2-(3,4-dihydroxyphenyl)-5,7-dihydroxy-1lambda4-chromen-1-ylium"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "average_mass"
#> 
#> $property$value
#> [1] "491.4215"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "mono_mass"
#> 
#> $property$value
#> [1] "491.1189512"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "smiles"
#> 
#> $property$value
#> [1] "[H][C@]1(COC(C)=O)O[C@@]([H])(OC2=CC3=C(O)C=C(O)C=C3[O+]=C2C2=CC=C(O)C(O)=C2)[C@]([H])(O)[C@@]([H])(O)[C@@]1([H])O"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "formula"
#> 
#> $property$value
#> [1] "C23H23O12"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "inchi"
#> 
#> $property$value
#> [1] "InChI=1S/C23H22O12/c1-9(24)32-8-18-19(29)20(30)21(31)23(35-18)34-17-7-12-14(27)5-11(25)6-16(12)33-22(17)10-2-3-13(26)15(28)4-10/h2-7,18-21,23,29-31H,8H2,1H3,(H3-,25,26,27,28)/p+1/t18-,19+,20+,21-,23-/m1/s1"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "inchikey"
#> 
#> $property$value
#> [1] "HBXXDBKJLPLXPR-DLBZZEGUSA-O"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "polar_surface_area"
#> 
#> $property$value
#> [1] "199.51"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "refractivity"
#> 
#> $property$value
#> [1] "125.41"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "polarizability"
#> 
#> $property$value
#> [1] "47.4"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "rotatable_bond_count"
#> 
#> $property$value
#> [1] "6"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "acceptor_count"
#> 
#> $property$value
#> [1] "11"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "donor_count"
#> 
#> $property$value
#> [1] "7"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "physiological_charge"
#> 
#> $property$value
#> [1] "0"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $property
#> $property$kind
#> [1] "formal_charge"
#> 
#> $property$value
#> [1] "1"
#> 
#> $property$source
#> [1] "ChemAxon"
#> 
#> 
#> $pathways
#> NULL
#> 
#> $spectra
#>             spectrum      spectrum      spectrum      spectrum     
#> type        "Specdb::CMs" "Specdb::CMs" "Specdb::CMs" "Specdb::CMs"
#> spectrum_id "23123"       "41216"       "137797"      "145531"     
#>             spectrum       spectrum       spectrum       spectrum      
#> type        "Specdb::MsMs" "Specdb::MsMs" "Specdb::MsMs" "Specdb::MsMs"
#> spectrum_id "81423"        "81424"        "81425"        "142524"      
#>             spectrum       spectrum       spectrum       spectrum      
#> type        "Specdb::MsMs" "Specdb::MsMs" "Specdb::MsMs" "Specdb::MsMs"
#> spectrum_id "142525"       "142526"       "2851882"      "2851883"     
#>             spectrum      
#> type        "Specdb::MsMs"
#> spectrum_id "2851884"     
#> 
#> $hmdb_id
#> [1] "HMDB29236"
#> 
#> $pubchem_compound_id
#> NULL
#> 
#> $chemspider_id
#> NULL
#> 
#> $kegg_id
#> NULL
#> 
#> $chebi_id
#> NULL
#> 
#> $biocyc_id
#> NULL
#> 
#> $het_id
#> NULL
#> 
#> $wikipidia
#> NULL
#> 
#> $vmh_id
#> NULL
#> 
#> $fbonto_id
#> NULL
#> 
#> $foodb_id
#> NULL
#> 
#> $general_references
#> NULL
#> 
#> $foods
#>                  food                   food                     
#> name             "Highbush blueberry"   "Lowbush blueberry"      
#> food_type        "Type 1"               "Type 1"                 
#> category         "specific"             "specific"               
#> name_scientific  "Vaccinium corymbosum" "Vaccinium angustifolium"
#> ncbi_taxonomy_id "69266"                "472369"                 
#> average_value    "0.103886"             "0.664873"               
#> max_value        "0.103886414"          "0.664873055"            
#> min_value        "0.103886414"          "0.664873055"            
#> unit             "mg/100 g"             "mg/100 g"               
#> 
#> $flavors
#> NULL
#> 
#> $enzymes
#> NULL
#> 
#> $health_effects
#> NULL
#> 
y = request_foodb_compound( compound_id = "FDB000004", return_form = "data.frame")