Request one specific the metabolite information in HMDB

Request one specific the metabolite information in HMDB

request_hmdb_metabolite(
  url = "https://hmdb.ca/metabolites",
  metabolite_id = "HMDB0000001",
  return_form = c("list", "data.frame")
)

Arguments

url

Default is "https://hmdb.ca/metabolites".

metabolite_id

metabolite id. For example, HMDB0000001

return_form

data.frame or list.

Value

A data frame or list.

Author

Xiaotao Shen shenxt1990@outlook.com

Examples

x = request_hmdb_metabolite(metabolite_id = "HMDB0000001", return_form = "list")
x[1:2]
#> $version
#> [1] "5.0"
#> 
#> $creation_date
#> [1] "2005-11-16 15:48:42 UTC"
#> 
y = request_hmdb_metabolite(metabolite_id = "HMDB0000001", return_form = "data.frame")
head(y)
#>   version           creation_date             update_date   accession
#> 1     5.0 2005-11-16 15:48:42 UTC 2022-07-27 18:31:18 UTC HMDB0000001
#>                name
#> 1 1-Methylhistidine
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          description
#> 1 1-Methylhistidine, also known as 1-MHis, 1MH, tau-methylhistidine or tele-methylhistidine, belongs to the class of organic compounds known as histidine and derivatives. 1MH is also classified as a methylamino acid. Methylamino acids are primarily proteogenic amino acids (found in proteins) which have been methylated (in situ) on their side chains by various methyltransferase enzymes. Histidine can be methylated at either the N1 or N3 position of its imidazole ring, yielding the isomers 1-methylhistidine (1MH; also referred to as tau-methylhistidine, according to IUPAC) or 3-methylhistidine (3MH; pi-methylhistidine, according to IUPAC), respectively. There is considerable confusion with regard to the nomenclature of the methylated nitrogen atoms on the imidazole ring of histidine in histidine-containing proteins (such as actin and myosin) as well as histidine-containing peptides (such as anserine and ophidine/balenine). In particular, older literature (mostly prior to the year 2000) as well as most biochemists and nutrition scientists incorrectly number the imidazole nitrogen atom most proximal to the side chain beta-carbon as 1 or N1, while organic chemists correctly designate it as 3 or N3. As a result, biochemists and nutrition scientists historically designated anserine (Npi-methylated) as beta-alanyl-N1-methylhistidine (or beta-alanyl-1-methylhistidine), whereas according to standard IUPAC nomenclature, anserine is correctly named as beta-alanyl-N3-methylhistidine. As a result, for several decades, many papers incorrectly identified 1MH as a specific marker for dietary consumption or various pathophysiological effects when they really are referring to 3MH – and vice versa (PMID: 24137022). To help resolve this issue the IUPAC commission (PMID: 6743224 and IUPAC Compendium of Chemical Terminology, 2nd ed. (the 'Gold Book'). Compiled by A. D. McNaught and A. Wilkinson. Blackwell Scientific Publications, Oxford (1997)) revised the nomenclature for histidine and introduced the terms pi (for prox or pros – near) and tau (for tele – far) to label the imidazole nitrogens in histidine. Therefore, the pi nitrogen is the nitrogen closest to the side chain beta carbon (atom #3 or N3) while the tau nitrogen is most distant from the side chain beta carbon (atom #1 or N1). IUPAC's goal is for the global community to refer to the molecule depicted here is as 'tau-methylhistidine' with the hope that the archaic term, 1-methylhistidine will eventually disappear. Unfortunately, this has not happened and confusion still persists.  Older versions of the HMDB (prior to 2022) as well as current versions of some databases, such as PubChem, KEGG, and UniProt, indicate that an acceptable synonym for 1MH is pi-methylhistidine or otherwise somehow equate 1MH and 3MH. This is incorrect and it continues to sow confusion. Indeed, a key paper that identified METTL9 as the enzyme responsible for pi-methylation of histidine in most vertebrates also incorrectly labeled the METTL9 product as 1MH (PMID: 33563959). Similarly, a key paper that identified METTL18 as the enzyme responsible for tau-methylation of histidine incorrectly labelled the METTL18 product as 3MH (PMID: 33693809). Likewise, many members of the biochemical/nutrition community still incorrectly refer to 1MH as pi-methlyhistidine and 3MH as tau-methylhistidine. This has led to even more confusion. To maintain consistency for this compound description, all papers cited herein that incorrectly refer to 3MH as 1MH and vice versa, will have their conclusions re-stated and the citation will be marked with the phrase '3MH/1MH switch'. 1MH is a free amino acid arising from the proteolysis of 1MH-containing proteins and peptides.  It is not synthesized on its own, nor can it be incorporated into proteins as an amino acid. However, it can be incorporated into certain dipeptides through the action of the enzyme known as carnosine synthase I. 1MH can only be generated from histidine residues through the action of methyltransferases as a protein post-translational modification event. Histidine methylation on the 1- or tau site of histidine-containing proteins is mediated by at least two enzymes: SETD3 (PMID: 30526847) and METTL18 (3MH/1MH switch - PMID: 33693809).  SETD3, or SET domain-containing protein 3, is a protein-histidine N-methyltransferase that specifically mediates 1-methylhistidine (tau-methylhistidine) methylation of actin at 'His-73' (PMID: 30526847; 3MH/1MH switch - PMID: 30626964). SETD3 is a methyltransferase that uses S-adenosyl-L-methionine to transfer the methyl group to histidine at the tau position. Histidine methylation of actin His-73 is required for smooth muscle contraction of the laboring uterus during delivery (3MH/1MH switch - PMID: 30626964). It also reduces the nucleotide exchange rate on actin monomers and modestly accelerates actin filament assembly (3MH/1MH switch - PMID: 30626964). SETD3-mediated histidine methylation appears to occur in all higher eukaryotes with actin, from plants to insects to vertebrates. Within cells, SETD3 is found in the cytoplasm and nucleus. In contrast to SETD3, METTL18 is a nuclear methyltransferase protein that contains a functional nuclear localization signal and accumulates in nucleoli. Specifically, METTL18 is a seven β-strand (7BS) methyltransferase that uses S-adenosyl-L-methionine to transfer the methyl group to the tau position of His-245 on ribosomal protein L3, RPL3 (3MH/1MH switch - PMID: 33693809).  METTL18 is highly conserved and found in essentially all eukaryotes (from yeast to humans). METTL18-mediated methylation of RPL3 is important for optimal ribosome biogenesis and function (3MH/1MH switch - PMID: 33693809). Other proteins that are known to have 1MH modifications include myosin and myosin kinase.  In addition to these tau-His-methylated proteins, a specialized dipeptide called ophidine (balenine) that consists of beta-alanine and 1MH is also known (PMID: 24137022). This methylated analog of carnosine, which is naturally produced in the liver via the enzyme carnosine synthase I (PMID: 20097752), is especially abundant in the skeletal muscles and brains of whales and dolphins but almost completely absent in other vertebrates (PMID: 24137022).  Ophidine, like its homologs anserine and carnosine, is believed to act as a pH buffer (for lactic acid generated by muscles), an antiglycating agent, and an antioxidant. Neither ophidine nor anserine are produced in humans, with humans being the only vertebrate not producing (or producing very little) methylated histidine versions of carnosine (PMID: 24137022).  Because 1MH is so abundant in skeletal muscle tissues (being found in the main myofibrillar proteins actin and myosin), the urinary concentrations of 1-methylhistidine can be used as a biomarker for skeletal muscle protein breakdown, especially for those who have been subject to muscle injury (3MH/1MH switch - PMID: 16079625). During protein catabolism, 1-methylhistidine is released but cannot be reutilized. Therefore, the plasma concentration and urine excretion of 1-methylhistidine serve as sensitive markers of myofibrillar protein degradation (3MH/1MH switch - PMID: 32235743). Approximately 75% of 1-methylhistidine in the human body is estimated to originate from skeletal muscle (3MH/1MH switch - PMID: 32235743). In addition to the degradation of muscle proteins, the 1-methylhistidine level can be moderately affected by the degradation of intestinal proteins and meat intake. 1-Methylhistidine has been found to be associated with several diseases such as Alzheimer's disease, preeclampsia, obesity, kidney disease. The normal concentration of 1-methylhistidine in the urine of healthy adult humans has been detected and quantified in a range of 17.7-153.8 micromoles per millimole (umol/mmol) of creatinine, with most studies reporting the average urinary concentration between 25-40 umol/mmol of creatinine. The average concentration of 1-methylhistidine in human blood plasma has been detected and quantified at 12.7 micromolar (uM) with a range of 9.8-15.6 uM. As a general rule, urinary 3MH is associated with white meat intake (p< 0.001), whereas urinary 1MH is associated with red meat intake (p< 0.001) (3MH/1MH switch - PMID: 34091671).
#>                                                                                                                                                                                                                                                                                                                                                                                                                        synonyms
#> 1 (2S)-2-Amino-3-(1-methyl-1H-imidazol-4-yl)propanoic acid{}(2S)-2-Amino-3-(1-methyl-1H-imidazol-4-yl)propanoate{}1 Methylhistidine{}1-Methyl histidine{}1-Methyl-histidine{}1-Methyl-L-histidine{}1-MHis{}1-N-Methyl-L-histidine{}L-1-Methylhistidine{}N1-Methyl-L-histidine{}1-Methylhistidine dihydrochloride{}1-Methylhistidine{}Tau methylhistidine{}N(Tau)-methylhistidine{}Tau-methyl-L-histidine{}N Tau-methylhistidine
#>   chemical_formula average_molecular_weight monisotopic_moleculate_weight
#> 1         169.1811                 169.1811                            NA
#>   iupac_name traditional_iupac cas_registry_number
#> 1         NA                NA            332-80-9
#>                                                                         smiles
#> 1 [H]OC(=O)[C@@]([H])(N([H])[H])C([H])([H])C1=C([H])N(C([H])=N1)C([H])([H])[H]
#>                                                                                     inchi
#> 1 InChI=1S/C7H11N3O2/c1-10-3-5(9-4-10)2-6(8)7(11)12/h3-4,6H,2,8H2,1H3,(H,11,12)/t6-/m0/s1
#>                      inchikey state pathways hmdb_id pubchem_compound_id
#> 1 BRMWTNUJHUMWMS-LURJTMIESA-N Solid       NA      NA               92105
#>   chemspider_id kegg_id chebi_id biocyc_id het_id wikipidia vmh_id fbonto_id
#> 1         83153  C01152    50599        NA     NA        NA     NA        NA
#>   hmdb_id.1 flavors enzymes
#> 1        NA      NA      NA