Request one specific the metabolite information in HMDB

Request one specific the metabolite information in HMDB

request_hmdb_metabolite(
  url = "https://hmdb.ca/metabolites",
  metabolite_id = "HMDB0000001",
  return_form = c("list", "data.frame")
)

Arguments

url

Default is "https://hmdb.ca/metabolites".

metabolite_id

metabolite id. For example, HMDB0000001

return_form

data.frame or list.

Value

A data frame or list.

Author

Xiaotao Shen shenxt1990@outlook.com

Examples

x = request_hmdb_metabolite(metabolite_id = "HMDB0000001", return_form = "list")
x[1:2]
#> $version
#> [1] "5.0"
#> 
#> $creation_date
#> [1] "2005-11-16 15:48:42 UTC"
#> 
y = request_hmdb_metabolite(metabolite_id = "HMDB0000001", return_form = "data.frame")
head(y)
#>   version           creation_date             update_date   accession
#> 1     5.0 2005-11-16 15:48:42 UTC 2022-07-12 16:50:58 UTC HMDB0000001
#>                name
#> 1 1-Methylhistidine
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   description
#> 1 1-Methylhistidine, also known as 1-MHis or 1MH, belongs to the class of organic compounds known as histidine and derivatives. 1MH is also classified as a methylamino acid. Methylamino acids are primarily proteogenic amino acids (found in proteins) which have been methylated (in situ) on their side chains by various methyltransferase enzymes. Histidine can be methylated at either the N1 or N3 position of its imidazole ring, yielding the isomers 1-methylhistidine (1MH; also referred to as pi-methylhistidine) or 3-methylhistidine (3MH; tau-methylhistidine), respectively. There is considerable confusion with regard to the nomenclature of the methylated nitrogen atoms on the imidazole ring of histidine and other histidine-containing peptides such as anserine. In particular, older literature (mostly prior to the year 2000) designated anserine (Npi methylated) as beta-alanyl-N1-methyl-histidine, whereas according to standard IUPAC nomenclature, anserine is correctly named as beta-alanyl-N3-methyl-histidine. As a result, many papers published prior to the year 2000 incorrectly identified 1MH as a specific marker for dietary consumption or various pathophysiological effects when they really were referring to 3MH (PMID: 24137022).  Recent discoveries have shown that 1MH is produced in essentially all mammals (and other vertebrates) via the enzyme known as METTL9 (PMID: 33563959). METTL9 is a broad-specificity methyltransferase that mediates the formation of the majority of 1MH present in mammalian proteomes. METTL9-catalyzed methylation requires a His-x-His (HxH) motif, where "x" is a small amino acid. This HxH motif is found in a number of abundant mammalian proteins such as ARMC6, S100A9, and NDUFB3 (PMID: 33563959). Because of its abundance in many muscle-related proteins, 1MH has been found to be a good biomarker for the consumption of meat (PMID: 21527577). Dietary studies have shown that poultry consumption (p-trend = 0.0006) and chicken consumption (p-trend = 0.0003) are associated with increased levels of 1MH in human plasma (PMID: 30018457). The consumption of fish, especially salmon and cod, has also been shown to increase the levels of 1MH in serum and urine (PMID: 31401679). As a general rule, urinary 1MH is associated with white meat intake (p< 0.001), whereas urinary 3MH is associated with red meat intake (p< 0.001) (PMID: 34091671).
#>                                                                                                                                                                                                                                                                                                                                                                                                                        synonyms
#> 1 (2S)-2-Amino-3-(1-methyl-1H-imidazol-4-yl)propanoic acid{}(2S)-2-Amino-3-(1-methyl-1H-imidazol-4-yl)propanoate{}1 Methylhistidine{}1-Methyl histidine{}1-Methyl-histidine{}1-Methyl-L-histidine{}1-MHis{}1-N-Methyl-L-histidine{}L-1-Methylhistidine{}N1-Methyl-L-histidine{}1-Methylhistidine dihydrochloride{}1-Methylhistidine{}Tau methylhistidine{}N(Tau)-methylhistidine{}Tau-methyl-L-histidine{}N Tau-methylhistidine
#>   chemical_formula average_molecular_weight monisotopic_moleculate_weight
#> 1         169.1811                 169.1811                            NA
#>   iupac_name traditional_iupac cas_registry_number
#> 1         NA                NA            332-80-9
#>                                                                         smiles
#> 1 [H]OC(=O)[C@@]([H])(N([H])[H])C([H])([H])C1=C([H])N(C([H])=N1)C([H])([H])[H]
#>                                                                                     inchi
#> 1 InChI=1S/C7H11N3O2/c1-10-3-5(9-4-10)2-6(8)7(11)12/h3-4,6H,2,8H2,1H3,(H,11,12)/t6-/m0/s1
#>                      inchikey state pathways hmdb_id pubchem_compound_id
#> 1 BRMWTNUJHUMWMS-LURJTMIESA-N Solid       NA      NA               92105
#>   chemspider_id kegg_id chebi_id biocyc_id het_id wikipidia vmh_id fbonto_id
#> 1         83153  C01152    50599        NA     NA        NA     NA        NA
#>   hmdb_id.1 flavors enzymes
#> 1        NA      NA      NA